test: UserServiceTest를 기능 단위로 분리
- 비대해진 UserServiceTest를 사용자 목록 조회, 정보 수정, 권한 수정 테스트로 나눈다.
- UserServiceListTest에서 목록 조회 권한, 검색, 필터, 페이지네이션, 정렬 검증을 담당한다.
- UserServiceUpdateTest에서 본인 정보 수정과 관리자 사용자 정보 수정 검증을 담당한다.
- UserServiceRoleTest에서 사용자 권한 수정과 MASTER 권한 정책 검증을 담당한다.
- 중복되던 사용자 생성, 요청 DTO 생성, Principal 생성, 예외 검증 로직을 UserServiceTestFixture로 추출한다.
- 테스트 클래스별 책임을 분리해 가독성과 유지보수성을 개선한다.
